requirements.txt
# Core

fastapi==0.109.2
uvicorn[standard]==0.27.1

# Streamlit Core + Extensões

streamlit>=1.28.0
# streamlit==1.32.2
streamlit-autorefresh==1.0.1
# streamlit-toggle==0.1.3



openai==1.12.0
python-dotenv==1.0.1

# Data & Analysis
pandas==2.2.0
numpy==1.26.4

# Visualization
wordcloud==1.9.3
networkx==3.2.1
pyvis==0.3.2
matplotlib==3.8.2

# Database
redis==5.0.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Image
Pillow==10.2.0

# Utilities
python-multipart==0.0.7
pydantic==2.6.1
httpx==0.26.0

# Database
psycopg2-binary==2.9.9
neo4j>=5.0.0
#redis==5.0.1

